move InputSelectionWrapper to /algo_params dir, also simplify
